First install
A buddy of mine who is an investment banking analyst and I were
complaining about some of the more tedious parts of our jobs one day.
He said that every day he has to read multiple reports that can be up
to several hundred pages long, and he really wished that there was a
way for him to get a sense of what each one is about without reading
them thoroughly so he could prioritize them. He showed me a document he
created in Excel that could detect word frequency if a text were loaded
in. I thought this was kind of a clunky solution so I developed this
(with some help from users at StackOverflow: http://bit.ly/1JQJTbe).
This script create a function that will automatically read any .txt
file in the working directory and turn it into a Word Cloud. Simply run
the function passing the name of your file without its extension as an
argument (this is mislabelled in the script notes itself). In the
future I’m looking to build in support for .pdfs, but for now just as a
demonstration of concept I thought this was kinda neat. Enjoy!